Refactoring part two - Optimized AlertModelBuilder - Moved makeResultsMessage to Presenter
